HBX-1232: Add gradle support   - Pull up the management of the Hibernate properties to class 'org.hibernate.tool.gradle.task.AbstractTask'   - Remove the management of Hibernate properties from 'GenerateJavaTask' and 'RunSqlTask'
